+++ Packages/com.zebugger.zebug/CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## Unreleased
 
 ### Added
-* **WIP** in-zebug window graph.
-    * E.G. `Zebug.GraphValue(Mathf.Sin(Time.time))`  
+* Zebug editor window graph for visualizing value-over-time data.
+    * Usage: `Zebug.GraphValue(Mathf.Sin(Time.time))`  
+    * Added Mouse hover for nearest-data-sample data inspection.
 
 ### Removed
 * Removed preprocessor list functionality, wasn't really used.